requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
pydantic==2.11.7
pydantic-settings==2.10.1
pytest==8.4.1
pytest-asyncio==0.24.0
aiosqlite==0.20.0
httpx==0.28.1
python-dotenv==1.1.1
sqlalchemy==2.0.41
alembic==1.16.2
black==25.1.0
isort==6.0.1
flake8==7.3.0
mypy==1.16.1
pre-commit==4.2.0
resend==2.10.0
fastapi-sso==0.18.0
bcrypt==4.3.0
python-multipart==0.0.20
email-validator==2.2.0
PyJWT==2.10.1
asyncpg==0.29.0
openai==1.91.0
langchain==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.25
langchain-text-splitters==0.3.8
supabase==2.16.0
Pillow==11.2.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1